requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
ulid-py==1.1.0
jinja2==3.1.4
pytest==8.3.0
pytest-asyncio==0.24.0
